refactor: extract Infrastructure scan/probe helpers under 500 lines
Extract Class: move asmdef reference-requirement recording out of
ThirdPartyToolMigrationAssemblyUsageScanState and move login-shell CLI
probe command/parsing out of CliInstallationDetector. No behavior change;
characterization and existing CliInstallationDetector tests stay green.
