+++ Packages/src/Editor/Infrastructure/CLI/CliShellInstallationProbe.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5adfd558b96847638990c9dcf3a51025
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
