+++ Packages/src/Editor/Infrastructure/ThirdPartyToolMigration/ThirdPartyToolMigrationAssemblyUsageScanState.cs
@@ -131,33 +131,10 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
             }
 
             string assemblyDirectory = FindAssemblyDirectory(csharpFilePath);
-            string[] legacyAssemblyAliases = GetAssemblyScopedNames(
-                AssemblyScopedLegacyAliasesByDirectory,
+            ThirdPartyToolMigrationAssemblyUsageReferenceRequirementRecorder.Record(
+                this,
+                source,
                 assemblyDirectory);
-            string[] assemblyDeclaredTypeNames = GetAssemblyScopedNames(
-                AssemblyDeclaredTypeNamesByDirectory,
-                assemblyDirectory);
-            bool hasLegacyCSharpApi = ThirdPartyToolMigrationRules.ContainsLegacyCSharpApi(source);
-            RecordToolContractsRequirement(
-                source,
-                assemblyDirectory,
-                legacyAssemblyAliases,
-                assemblyDeclaredTypeNames,
-                hasLegacyCSharpApi);
-            RecordRegistrarDomainReturnRequirement(source, assemblyDirectory, legacyAssemblyAliases);
-            RecordDomainContractRequirement(source, assemblyDirectory);
-            RecordFirstPartyScreenshotRequirement(
-                source,
-                assemblyDirectory,
-                legacyAssemblyAliases,
-                assemblyDeclaredTypeNames,
-                hasLegacyCSharpApi);
-            RecordEditorWindowCaptureTimeoutRequirement(
-                source,
-                assemblyDirectory,
-                legacyAssemblyAliases,
-                assemblyDeclaredTypeNames,
-                hasLegacyCSharpApi);
         }
 
         public MigrationAssemblyUsage CreateUsage()
@@ -351,196 +328,6 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
                 DomainReferenceAssemblyDirectories.Add(assemblyDirectory);
                 ToolContractsReferenceAssemblyDirectories.Add(assemblyDirectory);
             }
-        }
-
-        private void RecordToolContractsRequirement(
-            string source,
-            string assemblyDirectory,
-            string[] legacyAssemblyAliases,
-            string[] assemblyDeclaredTypeNames,
-            bool hasLegacyCSharpApi)
-        {
-            string[] currentApplicationAssemblyAliases = GetAssemblyScopedNames(
-                AssemblyScopedCurrentApplicationAliasesByDirectory,
-                assemblyDirectory);
-            bool hasCurrentApplicationSourceTarget =
-                ThirdPartyToolMigrationRules.ContainsCurrentApplicationApiForAssembly(
-                    source,
-                    AssemblyScopedCurrentApplicationDirectories.Contains(assemblyDirectory),
-                    currentApplicationAssemblyAliases,
-                    assemblyDeclaredTypeNames);
-            if (!ContainsToolContractsReferenceRequirement(
-                    source,
-                    assemblyDirectory,
-                    legacyAssemblyAliases,
-                    currentApplicationAssemblyAliases,
-                    assemblyDeclaredTypeNames,
-                    hasLegacyCSharpApi,
-                    hasCurrentApplicationSourceTarget))
-            {
-                return;
-            }
-
-            ToolContractsReferenceAssemblyDirectories.Add(assemblyDirectory);
-        }
-
-        private bool ContainsToolContractsReferenceRequirement(
-            string source,
-            string assemblyDirectory,
-            string[] legacyAssemblyAliases,
-            string[] currentApplicationAssemblyAliases,
-            string[] assemblyDeclaredTypeNames,
-            bool hasLegacyCSharpApi,
-            bool hasCurrentApplicationSourceTarget)
-        {
-            bool canUseLegacyAssemblyApi =
-                AssemblyScopedLegacyDirectories.Contains(assemblyDirectory) ||
-                hasLegacyCSharpApi ||
-                ThirdPartyToolMigrationRules.ContainsCurrentToolContractsApi(source) ||
-                AssemblyScopedCurrentToolContractsDirectories.Contains(assemblyDirectory) ||
-                AssemblyScopedCurrentApplicationDirectories.Contains(assemblyDirectory);
-            return ThirdPartyToolMigrationRules.ContainsLegacyRegistrarApiForAssembly(
-                    source,
-                    AssemblyScopedLegacyDirectories.Contains(assemblyDirectory),
-                    legacyAssemblyAliases) ||
-                ThirdPartyToolMigrationRules.ContainsLegacyApplicationApiForAssembly(
-                    source,
-                    canUseLegacyAssemblyApi,
-                    legacyAssemblyAliases,
-                    currentApplicationAssemblyAliases,
-                    assemblyDeclaredTypeNames) ||
-                hasCurrentApplicationSourceTarget;
-        }
-
-        private void RecordRegistrarDomainReturnRequirement(
-            string source,
-            string assemblyDirectory,
-            string[] legacyAssemblyAliases)
-        {
-            if (!ThirdPartyToolMigrationRules.ContainsRegistrarDomainReturnApiForAssembly(
-                    source,
-                    AssemblyScopedLegacyDirectories.Contains(assemblyDirectory),
-                    legacyAssemblyAliases))
-            {
-                return;
-            }
-
-            ToolContractsReferenceAssemblyDirectories.Add(assemblyDirectory);
-        }
-
-        private void RecordDomainContractRequirement(string source, string assemblyDirectory)
-        {
-            string[] currentDomainAssemblyAliases = GetAssemblyScopedNames(
-                AssemblyScopedCurrentDomainAliasesByDirectory,
-                assemblyDirectory);
-            string[] currentDomainNamespaceAliases =
-                ThirdPartyToolMigrationAliasRules.GetCombinedCurrentDomainNamespaceAliases(
-                    source,
-                    currentDomainAssemblyAliases);
-            if (!ThirdPartyToolMigrationRules.ContainsCurrentDomainMetadataApiForAssembly(
-                    source,
-                    AssemblyScopedCurrentDomainDirectories.Contains(assemblyDirectory)) &&
-                !ThirdPartyToolMigrationRules.ContainsCurrentDomainContractAliasReference(
-                    source,
-                    currentDomainNamespaceAliases))
-            {
-                return;
-            }
-
-            ToolContractsReferenceAssemblyDirectories.Add(assemblyDirectory);
-        }
-
-        private void RecordFirstPartyScreenshotRequirement(
-            string source,
-            string assemblyDirectory,
-            string[] legacyAssemblyAliases,
-            string[] assemblyDeclaredTypeNames,
-            bool hasLegacyCSharpApi)
-        {
-            string[] currentFirstPartyToolsAssemblyAliases = GetAssemblyScopedNames(
-                AssemblyScopedCurrentFirstPartyToolsAliasesByDirectory,
-                assemblyDirectory);
-            bool hasCurrentFirstPartyScreenshotReferenceRequirement =
-                ThirdPartyToolMigrationRules.ContainsCurrentFirstPartyScreenshotApiForAssembly(
-                    source,
-                    AssemblyScopedCurrentFirstPartyToolsDirectories.Contains(assemblyDirectory),
-                    currentFirstPartyToolsAssemblyAliases,
-                    assemblyDeclaredTypeNames);
-            if (ContainsFirstPartyScreenshotToolContractsRequirement(
-                source,
-                assemblyDirectory,
-                legacyAssemblyAliases,
-                currentFirstPartyToolsAssemblyAliases,
-                assemblyDeclaredTypeNames,
-                hasLegacyCSharpApi,
-                hasCurrentFirstPartyScreenshotReferenceRequirement))
-            {
-                ToolContractsReferenceAssemblyDirectories.Add(assemblyDirectory);
-            }
-
-            if (hasCurrentFirstPartyScreenshotReferenceRequirement)
-            {
-                FirstPartyScreenshotReferenceAssemblyDirectories.Add(assemblyDirectory);
-            }
-        }
-
-        private bool ContainsFirstPartyScreenshotToolContractsRequirement(
-            string source,
-            string assemblyDirectory,
-            string[] legacyAssemblyAliases,
-            string[] currentFirstPartyToolsAssemblyAliases,
-            string[] assemblyDeclaredTypeNames,
-            bool hasLegacyCSharpApi,
-            bool hasCurrentFirstPartyScreenshotReferenceRequirement)
-        {
-            bool hasAssemblyScopedCurrentToolContractsUsing =
-                AssemblyScopedCurrentToolContractsDirectories.Contains(assemblyDirectory);
-            bool hasAssemblyScopedCurrentFirstPartyToolsUsing =
-                AssemblyScopedCurrentFirstPartyToolsDirectories.Contains(assemblyDirectory);
-            return ThirdPartyToolMigrationRules.ContainsLegacyFirstPartyScreenshotApiForAssembly(
-                    source,
-                    AssemblyScopedLegacyDirectories.Contains(assemblyDirectory) ||
-                    hasLegacyCSharpApi ||
-                    ThirdPartyToolMigrationRules.ContainsCurrentToolContractsApi(source) ||
-                    hasAssemblyScopedCurrentToolContractsUsing,
-                    legacyAssemblyAliases,
-                    assemblyDeclaredTypeNames) ||
-                ThirdPartyToolMigrationRules.ContainsLegacyEditorWindowCaptureUtilityMigrationForAssembly(
-                    source,
-                    AssemblyScopedLegacyDirectories.Contains(assemblyDirectory) || hasLegacyCSharpApi,
-                    hasAssemblyScopedCurrentToolContractsUsing,
-                    hasAssemblyScopedCurrentFirstPartyToolsUsing,
-                    legacyAssemblyAliases,
-                    currentFirstPartyToolsAssemblyAliases,
-                    assemblyDeclaredTypeNames) ||
-                ThirdPartyToolMigrationRules.ContainsCurrentFirstPartyScreenshotContractApiForAssembly(
-                    source,
-                    hasAssemblyScopedCurrentFirstPartyToolsUsing,
-                    currentFirstPartyToolsAssemblyAliases,
-                    assemblyDeclaredTypeNames) ||
-                hasCurrentFirstPartyScreenshotReferenceRequirement;
-        }
-
-        private void RecordEditorWindowCaptureTimeoutRequirement(
-            string source,
-            string assemblyDirectory,
-            string[] legacyAssemblyAliases,
-            string[] assemblyDeclaredTypeNames,
-            bool hasLegacyCSharpApi)
-        {
-            if (!ThirdPartyToolMigrationRules.ContainsLegacyEditorWindowCaptureUtilityTimeoutMigrationForAssembly(
-                    source,
-                    AssemblyScopedLegacyDirectories.Contains(assemblyDirectory) || hasLegacyCSharpApi,
-                    AssemblyScopedCurrentToolContractsDirectories.Contains(assemblyDirectory),
-                    AssemblyScopedCurrentFirstPartyToolsDirectories.Contains(assemblyDirectory),
-                    legacyAssemblyAliases,
-                    GetAssemblyScopedNames(AssemblyScopedCurrentFirstPartyToolsAliasesByDirectory, assemblyDirectory),
-                    assemblyDeclaredTypeNames))
-            {
-                return;
-            }
-
-            ToolContractsReferenceAssemblyDirectories.Add(assemblyDirectory);
         }
 
         private void AddRegistrarApplicationReferences()
